Move the remaining methods from Base to Core...it actually seems to work.
